Bug 678 (fix), Bug 641 (API + Windows Impl.), Bug 688 (prep): Update NEWT's KeyEvent handling while distinguish keyCode (kbd layout independent) and keySym (kbd layout dependent)
API Changes:

- Virtual key codes and symbols are of type short.

- KeyEvent.keySymbol() shall return a layout dependent value (Bug 641)
  - Method returns former keyCode() value, which was layout dependent.
  - Returns 'short' value

- KeyEvent.keyCode() returns a short value, instead of int

- KeyEvent.keyCode() shall return a layout independent value (Bug 641)
  - To ease implementation, we only 'require' the scan code to be mapped to a 'US Keyboard layout',
    which allows reusing layout dependent code while preserving the goal to have a fixed physical key association

  - Implementation status:
    - Windows OK
    - X11 TODO
    - OSX: 50/50 TODO
      - Using layout independent 'action keys'
      - Using layout dependent 'printable keys'
      - returning above semantics for both, keyCode and keySym

    - Android 50/50 TODO
      - Returning the layout independent keyCode
      - Mapping probably incomplete

- KeyEvent.EVENT_KEY_TYPED and KeyListener.keyTyped(KeyEvent) (Bug 688)
  - Marked DEPRECATED
  - No more called for auto-repeat events
  - Synthesized in WindowImpl.consumeKeyEvent(..): No more injection by native- or java driver code

- NEWTEvent.eventType: int -> short
  - field, as well as all method involving eventType changed to short.

- NEWTEvent.isSystemEvent: REMOVED
  - Never used as well as never being implemented properly

Internal Changes:

- Simplified keyEvent driver code
  - Especially the Windows native driver's mapping code
    could be simplified using scanCode and MapVirtualKeyEx

- NEWT Event Factories: hashMap -> switch/case

Unit Tests:
  -
  - Added NewtCanvasAWT Offscreen Layer Tests
    important to test the AWT -> NEWT translation on OSX/CALayer:
    - TestNewtKeyCodeModifiersAWT
    - TestNewtKeyCodesAWT
    - TestNewtKeyEventAutoRepeatAWT
    - TestNewtKeyEventOrderAWT
    - TestNewtKeyPressReleaseUnmaskRepeatAWT

+++ src/newt/classes/com/jogamp/newt/event/InputEvent.java
@@ -86,7 +86,7 @@ public abstract class InputEvent extends NEWTEvent
   */
  public static final Object consumedTag = new Object();
  
- protected InputEvent(int eventType, Object source, long when, int modifiers) {
+ protected InputEvent(short eventType, Object source, long when, int modifiers) {
     super(eventType, source, when);
     this.modifiers=modifiers;
  }
@@ -154,16 +154,16 @@ public abstract class InputEvent extends NEWTEvent
   * @return Array of pressed mouse buttons  [{@link MouseEvent#BUTTON1} .. {@link MouseEvent#BUTTON6}]. 
   *         If none is down, the resulting array is of length 0.
   */
- public final int[] getButtonsDown()  {
+ public final short[] getButtonsDown()  {
      int len = 0;
      for(int i=1; i<=MouseEvent.BUTTON_NUMBER; i++) {
-         if(isButtonDown(i)) { len++; }
+         if( isButtonDown(i) ) { len++; }
      }
      
-     int[] res = new int[len];
+     short[] res = new short[len];
      int j = 0;
      for(int i=1; i<=MouseEvent.BUTTON_NUMBER; i++) {
-         if(isButtonDown(i)) { res[j++] = ( MouseEvent.BUTTON1 - 1 ) + i; }
+         if( isButtonDown(i) ) { res[j++] = (short) ( ( MouseEvent.BUTTON1 - 1 ) + i ); }
      }
      return res;
  }
