+++ src/newt/classes/com/jogamp/newt/event/WindowEvent.java
@@ -41,19 +41,19 @@ package com.jogamp.newt.event;
  */
 @SuppressWarnings("serial")
 public class WindowEvent extends NEWTEvent {
-    public static final int EVENT_WINDOW_RESIZED = 100;
-    public static final int EVENT_WINDOW_MOVED   = 101; 
-    public static final int EVENT_WINDOW_DESTROY_NOTIFY = 102;
-    public static final int EVENT_WINDOW_GAINED_FOCUS = 103;
-    public static final int EVENT_WINDOW_LOST_FOCUS = 104;
-    public static final int EVENT_WINDOW_REPAINT = 105;
-    public static final int EVENT_WINDOW_DESTROYED = 106;
+    public static final short EVENT_WINDOW_RESIZED = 100;
+    public static final short EVENT_WINDOW_MOVED   = 101; 
+    public static final short EVENT_WINDOW_DESTROY_NOTIFY = 102;
+    public static final short EVENT_WINDOW_GAINED_FOCUS = 103;
+    public static final short EVENT_WINDOW_LOST_FOCUS = 104;
+    public static final short EVENT_WINDOW_REPAINT = 105;
+    public static final short EVENT_WINDOW_DESTROYED = 106;
 
-    public WindowEvent(int eventType, Object source, long when) {
+    public WindowEvent(short eventType, Object source, long when) {
         super(eventType, source, when);
     }
 
-    public static String getEventTypeString(int type) {
+    public static String getEventTypeString(short type) {
         switch(type) {
             case EVENT_WINDOW_RESIZED: return "WINDOW_RESIZED";
             case EVENT_WINDOW_MOVED: return "WINDOW_MOVED";
